refactor(worker): move tar logic to tar
This commit creates the Tar class to handle all the responsibilities
around making a Tar.

In order to test the creation of an actual .tar file using the Tar
class, an empty fixtures directory is now part of the repo. Tests that use
this empty fixtures directory to test File IO need to manage creation
and cleanup of any files needed. See the specs for Tar#make as an
example.