chore(Rakefile,circle.yml): add rubocop rake task
This commit also adds a :pre_commit Rake task and sets that task as the
default. This task will run the :rubocop and :rspec tasks that are run
by CircleCI. Hopefully this will help people to verify their commits
before pushing to CircleCI and getting a failure from one of these rake
tasks.